Rename package as 'top.gcszhn'
