+++ src/main/java/top/gcszhn/autocard/utils/DigestUtils.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright © 2021 <a href="mailto:zhang.h.n@foxmail.com">Zhang.H.N</a>.
+ *
+ * Licensed under the Apache License, Version 2.0 (thie "License");
+ * You may not use this file except in compliance with the license.
+ * You may obtain a copy of the License at
+ *
+ *       http://wwww.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language govering permissions and
+ * limitations under the License.
+ */
+package top.gcszhn.autocard.utils;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.Base64;
+
+
+public class DigestUtils {
+    public static String digest(String message, String algorithm) throws NoSuchAlgorithmException {
+        byte[] data = message.getBytes(StandardCharsets.UTF_8);
+        MessageDigest md = MessageDigest.getInstance(algorithm);
+        md.update(data);
+        return new String(Base64.getEncoder().encode(md.digest()), StandardCharsets.UTF_8);
+    }
+}
